backend/requirements.txt
asgiref
black
celery
Django
django-cors-headers
django-environ
django-filter
djangorestframework
flake8
gunicorn
isort
Pillow
redis
sqlparse
